requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
pydantic>=2.8.0
pydantic-settings>=2.3.0
transformers>=4.44.0
torch>=2.4.0
google-genai>=0.5.0
pandas>=2.2.0
scikit-learn>=1.5.0
google-genai>=0.5.0
